+++ src/spec/EntitySpec.ts
@@ -377,4 +377,53 @@ describe('An entity', () => {
     expect(grandchild.scene).toBe(null);
 
   });
+
+  it('can removeAllChildren correctly', () => {
+    const e = new ex.Entity();
+    const child1 = new ex.Entity();
+    const child2 = new ex.Entity();
+    const child3 = new ex.Entity();
+    const child4 = new ex.Entity();
+    const child5 = new ex.Entity();
+    const child6 = new ex.Entity();
+    e.addChild(child1);
+    e.addChild(child2);
+    e.addChild(child3);
+    e.addChild(child4);
+    e.addChild(child5);
+    e.addChild(child6);
+
+    expect(e.children.length).toBe(6);
+
+    e.removeAllChildren();
+
+    expect(e.children.length).toBe(0);
+  });
+
+  it('will remove children from the parent when a child is killed', () => {
+    const e = new ex.Entity();
+    const child1 = new ex.Entity();
+    const child2 = new ex.Entity();
+    const child3 = new ex.Entity();
+    const child4 = new ex.Entity();
+    const child5 = new ex.Entity();
+    const child6 = new ex.Entity();
+    e.addChild(child1);
+    e.addChild(child2);
+    e.addChild(child3);
+    e.addChild(child4);
+    e.addChild(child5);
+    e.addChild(child6);
+
+    expect(e.children.length).toBe(6);
+
+    child1.kill();
+    child2.kill();
+    child3.kill();
+    child4.kill();
+    child5.kill();
+    child6.kill();
+
+    expect(e.children.length).toBe(0);
+  });
 });